apps/agent-service/pyproject.toml
[project]
name = "agent-service"
version = "0.1.0"
description = "AI Service API"
requires-python = ">=3.11"
dependencies = [
    "aio-pika>=9.5.0",
    "aioredis>=2.0.1",
    "arrow>=1.3.0",
    "asyncpg>=0.30.0",
    "fastapi>=0.116.1",
    "google-genai>=1.0.0",
    "greenlet>=3.2.3",
    "jieba>=0.42.1",
    "jinja2>=3.1.6",
    "json-repair>=0.54.0",
    "langchain>=0.3.26",
    "langchain-google-genai>=2.0.0",
    "langchain-openai>=0.3.28",
    "langfuse>=3.3.4",
    "langgraph>=0.5.3",
    "markupsafe>=3.0.2",
    "mypy-extensions>=1.1.0",
    "numpy>=2.2.6",
    "ollama>=0.5.1",
    "openai>=1.97.0",
    "packaging>=25.0",
    "pathspec>=0.12.1",
    "platformdirs>=4.3.8",
    "pydantic-settings>=2.10.1",
    "python-json-logger>=3.3.0",
    "qdrant-client>=1.14.3",
    "redis>=6.2.0",
    "requests>=2.32.4",
    "setuptools>=80.9.0",
    "sqlalchemy>=2.0.41",
    "sse-starlette>=2.4.1",
    "starlette>=0.47.1",
    "tenacity>=9.1.2",
    "uvicorn>=0.35.0",
    "arq>=0.25.0",
    "pytz>=2025.2",
    "bidict>=0.23.1",
    "volcengine-python-sdk[ark]>=5.0.3",
    "inner-shared",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.11.0",
    "pre-commit>=4.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.uv.sources]
inner-shared = { path = "../../packages/py-shared", editable = true }


[dependency-groups]
dev = [
    "factory-boy>=3.3.3",
    "faker>=37.4.2",
    "freezegun>=1.5.3",
    "httpx>=0.28.1",
    "pre-commit>=4.2.0",
    "pytest>=8.4.1",
    "pytest-asyncio>=0.26.0",
    "pytest-cov>=6.2.1",
    "pytest-mock>=3.14.1",
    "pytest-xdist>=3.8.0",
    "ruff>=0.12.5",
]
